# AVX512 wide universal intrinsics