… modules/core/CMakeLists.txt
… modules/core/include/opencv2/core/hal/intrin_forward.hpp
… modules/core/test/test_intrin.cpp
… modules/core/test/test_intrin512.simd.hpp
… modules/core/test/test_intrin_utils.hpp
… modules/imgproc/src/contours.cpp
… modules/imgproc/src/resize.cpp
… modules/imgproc/src/sumpixels.cpp