relates #13891

Implementation of 512-bit wide universal intrinsics for AVX512
https://github.com/opencv/opencv/wiki/OE-29.-Adding-AVX512-Support

<cut/>

TODO:
- [x] extend test_intrin tests

```
force_builders=docs,linux,Custom
buildworker:Custom=linux-3
build_image:Custom=ubuntu:18.04
CPU_BASELINE:Custom=AVX512_SKX
build_image:Custom Win=msvs2019
CPU_BASELINE:Custom Win=AVX512_SKX
#build_image:Custom Mac=
CPU_BASELINE:Custom Mac=AVX512_SKX
```