@terfendail 2019-04-26
<details>

<summary>Preliminary performance for core</summary>

| Test | AVX2 baseline time | AVX512 baseline time | ratio |
|------------------|----------------|----------|---------|
|CustomPtr::OCL_UMatTest::(1920x1080, false, 4)|0.337|0.298|1.13|
|CustomPtr::OCL_UMatTest::(1920x1080, false, 64)|0.267|0.268|1.00|
|CustomPtr::OCL_UMatTest::(1920x1080, false, 4096)|0.300|0.265|1.13|
|CustomPtr::OCL_UMatTest::(1920x1080, true, 4)|0.330|0.295|1.12|
|CustomPtr::OCL_UMatTest::(1920x1080, true, 64)|0.298|0.266|1.12|
|CustomPtr::OCL_UMatTest::(1920x1080, true, 4096)|0.268|0.263|1.02|
|CustomPtr::OCL_UMatTest::(3840x2160, false, 4)|1.383|1.279|1.08|
|CustomPtr::OCL_UMatTest::(3840x2160, false, 64)|1.164|1.149|1.01|
|CustomPtr::OCL_UMatTest::(3840x2160, false, 4096)|1.254|1.134|1.11|
|CustomPtr::OCL_UMatTest::(3840x2160, true, 4)|1.375|1.288|1.07|
|CustomPtr::OCL_UMatTest::(3840x2160, true, 64)|1.260|1.148|1.10|
|CustomPtr::OCL_UMatTest::(3840x2160, true, 4096)|1.177|1.113|1.06|
|CvRound_Float::Size_MatType::(127x61, 32FC1)|0.004|0.004|0.98|
|CvRound_Float::Size_MatType::(127x61, 64FC1)|0.004|0.004|1.01|
|CvRound_Float::Size_MatType::(640x480, 32FC1)|0.158|0.156|1.01|
|CvRound_Float::Size_MatType::(640x480, 64FC1)|0.166|0.164|1.01|
|CvRound_Float::Size_MatType::(1280x720, 32FC1)|0.476|0.466|1.02|
|CvRound_Float::Size_MatType::(1280x720, 64FC1)|0.562|0.540|1.04|
|CvRound_Float::Size_MatType::(1920x1080, 32FC1)|1.177|1.183|0.99|
|CvRound_Float::Size_MatType::(1920x1080, 64FC1)|1.596|1.624|0.98|
|LUT::SizePrm::320x240|0.039|0.029|1.33|
|LUT::SizePrm::640x480|0.155|0.115|1.35|
|LUT::SizePrm::1920x1080|1.041|0.766|1.36|
|Mat_Clone::Size_MatType::(127x61, 8UC1)|0.000|0.000|1.18|
|Mat_Clone::Size_MatType::(127x61, 32FC1)|0.001|0.001|1.62|
|Mat_Clone::Size_MatType::(127x61, 8UC4)|0.001|0.001|1.03|
|Mat_Clone::Size_MatType::(640x480, 8UC1)|0.011|0.011|1.01|
|Mat_Clone::Size_MatType::(640x480, 32FC1)|0.071|0.072|0.99|
|Mat_Clone::Size_MatType::(640x480, 8UC4)|0.073|0.071|1.02|
|Mat_Clone::Size_MatType::(1280x720, 8UC1)|0.040|0.038|1.04|
|Mat_Clone::Size_MatType::(1280x720, 32FC1)|0.272|0.270|1.01|
|Mat_Clone::Size_MatType::(1280x720, 8UC4)|0.271|0.269|1.01|
|Mat_Clone::Size_MatType::(1920x1080, 8UC1)|0.151|0.150|1.01|
|Mat_Clone::Size_MatType::(1920x1080, 32FC1)|0.701|0.710|0.99|
|Mat_Clone::Size_MatType::(1920x1080, 8UC4)|0.696|0.701|0.99|
|Mat_Clone_Roi::Size_MatType::(127x61, 8UC1)|0.000|0.000|1.09|
|Mat_Clone_Roi::Size_MatType::(127x61, 32FC1)|0.001|0.001|1.02|
|Mat_Clone_Roi::Size_MatType::(127x61, 8UC4)|0.001|0.001|1.01|
|Mat_Clone_Roi::Size_MatType::(640x480, 8UC1)|0.005|0.005|1.02|
|Mat_Clone_Roi::Size_MatType::(640x480, 32FC1)|0.044|0.044|1.00|
|Mat_Clone_Roi::Size_MatType::(640x480, 8UC4)|0.045|0.043|1.04|
|Mat_Clone_Roi::Size_MatType::(1280x720, 8UC1)|0.025|0.022|1.13|
|Mat_Clone_Roi::Size_MatType::(1280x720, 32FC1)|0.155|0.155|1.00|
|Mat_Clone_Roi::Size_MatType::(1280x720, 8UC4)|0.154|0.153|1.01|
|Mat_Clone_Roi::Size_MatType::(1920x1080, 8UC1)|0.089|0.089|1.01|
|Mat_Clone_Roi::Size_MatType::(1920x1080, 32FC1)|0.387|0.390|0.99|
|Mat_Clone_Roi::Size_MatType::(1920x1080, 8UC4)|0.383|0.389|0.99|
|Mat_CopyToWithMask::Size_MatType::(127x61, 8UC1)|0.000|0.000|2.12|
|Mat_CopyToWithMask::Size_MatType::(127x61, 16UC1)|0.001|0.000|1.70|
|Mat_CopyToWithMask::Size_MatType::(127x61, 32SC1)|0.003|0.003|1.03|
|Mat_CopyToWithMask::Size_MatType::(127x61, 8UC2)|0.001|0.000|1.75|
|Mat_CopyToWithMask::Size_MatType::(127x61, 8UC3)|0.005|0.005|1.00|
|Mat_CopyToWithMask::Size_MatType::(127x61, 32FC4)|0.004|0.004|1.00|
|Mat_CopyToWithMask::Size_MatType::(1920x1080, 8UC1)|0.230|0.226|1.02|
|Mat_CopyToWithMask::Size_MatType::(1920x1080, 16UC1)|0.420|0.399|1.05|
|Mat_CopyToWithMask::Size_MatType::(1920x1080, 32SC1)|1.174|1.203|0.98|
|Mat_CopyToWithMask::Size_MatType::(1920x1080, 8UC2)|0.428|0.406|1.06|
|Mat_CopyToWithMask::Size_MatType::(1920x1080, 8UC3)|1.420|1.355|1.05|
|Mat_CopyToWithMask::Size_MatType::(1920x1080, 32FC4)|4.265|4.295|0.99|
|Mat_Eye::Size_MatType::(127x61, 8UC1)|0.001|0.001|0.99|
|Mat_Eye::Size_MatType::(127x61, 32FC1)|0.001|0.001|1.05|
|Mat_Eye::Size_MatType::(127x61, 8UC4)|0.001|0.001|0.93|
|Mat_Eye::Size_MatType::(640x480, 8UC1)|0.007|0.007|1.01|
|Mat_Eye::Size_MatType::(640x480, 32FC1)|0.027|0.027|0.99|
|Mat_Eye::Size_MatType::(640x480, 8UC4)|0.025|0.025|1.01|
|Mat_Eye::Size_MatType::(1280x720, 8UC1)|0.017|0.017|1.03|
|Mat_Eye::Size_MatType::(1280x720, 32FC1)|0.114|0.115|0.99|
|Mat_Eye::Size_MatType::(1280x720, 8UC4)|0.112|0.112|1.00|
|Mat_Eye::Size_MatType::(1920x1080, 8UC1)|0.055|0.054|1.01|
|Mat_Eye::Size_MatType::(1920x1080, 32FC1)|0.283|0.285|0.99|
|Mat_Eye::Size_MatType::(1920x1080, 8UC4)|0.285|0.283|1.01|
|Mat_SetToWithMask::Size_MatType::(127x61, 8UC1)|0.002|0.001|1.16|
|Mat_SetToWithMask::Size_MatType::(127x61, 8UC2)|0.001|0.001|1.25|
|Mat_SetToWithMask::Size_MatType::(640x480, 8UC1)|0.018|0.009|2.03|
|Mat_SetToWithMask::Size_MatType::(640x480, 8UC2)|0.029|0.018|1.60|
|Mat_SetToWithMask::Size_MatType::(1280x720, 8UC1)|0.068|0.065|1.04|
|Mat_SetToWithMask::Size_MatType::(1280x720, 8UC2)|0.107|0.104|1.03|
|Mat_SetToWithMask::Size_MatType::(1920x1080, 8UC1)|0.155|0.152|1.02|
|Mat_SetToWithMask::Size_MatType::(1920x1080, 8UC2)|0.240|0.230|1.05|
|Mat_Transform::Size_MatType::(127x61, 8UC3)|0.016|0.016|1.01|
|Mat_Transform::Size_MatType::(127x61, 8SC3)|0.033|0.033|1.00|
|Mat_Transform::Size_MatType::(127x61, 16UC3)|0.013|0.013|1.00|
|Mat_Transform::Size_MatType::(127x61, 16SC3)|0.029|0.029|1.00|
|Mat_Transform::Size_MatType::(127x61, 32SC3)|0.023|0.023|1.00|
|Mat_Transform::Size_MatType::(127x61, 32FC3)|0.009|0.009|1.00|
|Mat_Transform::Size_MatType::(127x61, 64FC3)|0.014|0.005|2.79|
|Mat_Transform::Size_MatType::(640x480, 8UC3)|0.616|0.618|1.00|
|Mat_Transform::Size_MatType::(640x480, 8SC3)|1.237|1.273|0.97|
|Mat_Transform::Size_MatType::(640x480, 16UC3)|0.524|0.526|1.00|
|Mat_Transform::Size_MatType::(640x480, 16SC3)|1.115|1.151|0.97|
|Mat_Transform::Size_MatType::(640x480, 32SC3)|0.882|0.913|0.97|
|Mat_Transform::Size_MatType::(640x480, 32FC3)|0.385|0.511|0.75|
|Mat_Transform::Size_MatType::(640x480, 64FC3)|0.808|0.654|1.24|
|Mat_Transform::Size_MatType::(1280x720, 8UC3)|1.849|1.841|1.00|
|Mat_Transform::Size_MatType::(1280x720, 8SC3)|3.710|3.813|0.97|
|Mat_Transform::Size_MatType::(1280x720, 16UC3)|1.587|1.596|0.99|
|Mat_Transform::Size_MatType::(1280x720, 16SC3)|3.355|3.466|0.97|
|Mat_Transform::Size_MatType::(1280x720, 32SC3)|2.731|2.800|0.98|
|Mat_Transform::Size_MatType::(1280x720, 32FC3)|1.770|1.430|1.24|
|Mat_Transform::Size_MatType::(1280x720, 64FC3)|3.049|2.718|1.12|
|Mat_Transform::Size_MatType::(1920x1080, 8UC3)|4.164|4.172|1.00|
|Mat_Transform::Size_MatType::(1920x1080, 8SC3)|8.384|8.546|0.98|
|Mat_Transform::Size_MatType::(1920x1080, 16UC3)|3.587|3.616|0.99|
|Mat_Transform::Size_MatType::(1920x1080, 16SC3)|7.789|7.789|1.00|
|Mat_Transform::Size_MatType::(1920x1080, 32SC3)|6.272|6.245|1.00|
|Mat_Transform::Size_MatType::(1920x1080, 32FC3)|3.529|4.301|0.82|
|Mat_Transform::Size_MatType::(1920x1080, 64FC3)|7.026|6.535|1.08|
|Mat_Zeros::Size_MatType::(127x61, 8UC1)|0.000|0.000|0.97|
|Mat_Zeros::Size_MatType::(127x61, 32FC1)|0.000|0.000|0.94|
|Mat_Zeros::Size_MatType::(127x61, 32FC3)|0.001|0.001|1.00|
|Mat_Zeros::Size_MatType::(127x61, 8UC4)|0.000|0.000|0.88|
|Mat_Zeros::Size_MatType::(640x480, 8UC1)|0.004|0.004|1.00|
|Mat_Zeros::Size_MatType::(640x480, 32FC1)|0.021|0.021|1.00|
|Mat_Zeros::Size_MatType::(640x480, 32FC3)|0.109|0.109|1.00|
|Mat_Zeros::Size_MatType::(640x480, 8UC4)|0.023|0.022|1.01|
|Mat_Zeros::Size_MatType::(1280x720, 8UC1)|0.014|0.014|1.02|
|Mat_Zeros::Size_MatType::(1280x720, 32FC1)|0.109|0.108|1.00|
|Mat_Zeros::Size_MatType::(1280x720, 32FC3)|0.387|0.387|1.00|
|Mat_Zeros::Size_MatType::(1280x720, 8UC4)|0.108|0.109|1.00|
|Mat_Zeros::Size_MatType::(1920x1080, 8UC1)|0.050|0.050|1.01|
|Mat_Zeros::Size_MatType::(1920x1080, 32FC1)|0.278|0.277|1.00|
|Mat_Zeros::Size_MatType::(1920x1080, 32FC3)|0.990|1.021|0.97|
|Mat_Zeros::Size_MatType::(1920x1080, 8UC4)|0.279|0.277|1.01|
|abs::Size_MatType::(127x61, 8SC1)|0.004|0.003|1.40|
|abs::Size_MatType::(127x61, 32SC1)|0.002|0.002|1.26|
|abs::Size_MatType::(127x61, 32FC1)|0.002|0.002|1.29|
|abs::Size_MatType::(127x61, 8SC4)|0.015|0.010|1.53|
|abs::Size_MatType::(640x480, 8SC1)|0.124|0.092|1.35|
|abs::Size_MatType::(640x480, 32SC1)|0.098|0.106|0.92|
|abs::Size_MatType::(640x480, 32FC1)|0.094|0.106|0.89|
|abs::Size_MatType::(640x480, 8SC4)|0.528|0.362|1.46|
|abs::Size_MatType::(1280x720, 8SC1)|0.394|0.259|1.52|
|abs::Size_MatType::(1280x720, 32SC1)|0.289|0.286|1.01|
|abs::Size_MatType::(1280x720, 32FC1)|0.283|0.286|0.99|
|abs::Size_MatType::(1280x720, 8SC4)|1.556|1.065|1.46|
|abs::Size_MatType::(1920x1080, 8SC1)|0.752|0.598|1.26|
|abs::Size_MatType::(1920x1080, 32SC1)|0.839|0.879|0.95|
|abs::Size_MatType::(1920x1080, 32FC1)|0.839|0.870|0.96|
|abs::Size_MatType::(1920x1080, 8SC4)|3.510|2.390|1.47|
|absdiff::BinaryOpTest::(640x480, 8UC1)|0.010|0.011|0.94|
|absdiff::BinaryOpTest::(640x480, 8SC1)|0.010|0.011|0.94|
|absdiff::BinaryOpTest::(640x480, 16SC1)|0.065|0.062|1.05|
|absdiff::BinaryOpTest::(640x480, 32SC1)|0.135|0.133|1.02|
|absdiff::BinaryOpTest::(640x480, 32FC1)|0.134|0.134|1.00|
|absdiff::BinaryOpTest::(640x480, 16SC2)|0.135|0.134|1.00|
|absdiff::BinaryOpTest::(640x480, 8UC3)|0.101|0.100|1.01|
|absdiff::BinaryOpTest::(640x480, 16SC3)|0.202|0.200|1.01|
|absdiff::BinaryOpTest::(640x480, 8UC4)|0.136|0.133|1.02|
|absdiff::BinaryOpTest::(640x480, 16SC4)|0.274|0.269|1.02|
|absdiff::BinaryOpTest::(1280x720, 8UC1)|0.101|0.099|1.02|
|absdiff::BinaryOpTest::(1280x720, 8SC1)|0.102|0.101|1.01|
|absdiff::BinaryOpTest::(1280x720, 16SC1)|0.203|0.201|1.01|
|absdiff::BinaryOpTest::(1280x720, 32SC1)|0.433|0.425|1.02|
|absdiff::BinaryOpTest::(1280x720, 32FC1)|0.433|0.447|0.97|
|absdiff::BinaryOpTest::(1280x720, 16SC2)|0.433|0.428|1.01|
|absdiff::BinaryOpTest::(1280x720, 8UC3)|0.310|0.306|1.01|
|absdiff::BinaryOpTest::(1280x720, 16SC3)|0.745|0.727|1.03|
|absdiff::BinaryOpTest::(1280x720, 8UC4)|0.432|0.427|1.01|
|absdiff::BinaryOpTest::(1280x720, 16SC4)|1.102|1.096|1.01|
|absdiff::BinaryOpTest::(1920x1080, 8UC1)|0.230|0.226|1.02|
|absdiff::BinaryOpTest::(1920x1080, 8SC1)|0.230|0.226|1.02|
|absdiff::BinaryOpTest::(1920x1080, 16SC1)|0.499|0.477|1.05|
|absdiff::BinaryOpTest::(1920x1080, 32SC1)|1.315|1.311|1.00|
|absdiff::BinaryOpTest::(1920x1080, 32FC1)|1.312|1.452|0.90|
|absdiff::BinaryOpTest::(1920x1080, 16SC2)|1.283|1.318|0.97|
|absdiff::BinaryOpTest::(1920x1080, 8UC3)|0.861|0.836|1.03|
|absdiff::BinaryOpTest::(1920x1080, 16SC3)|2.088|2.061|1.01|
|absdiff::BinaryOpTest::(1920x1080, 8UC4)|1.300|1.319|0.99|
|absdiff::BinaryOpTest::(1920x1080, 16SC4)|2.880|2.883|1.00|
|absdiffScalarDouble::BinaryOpTest::(640x480, 8UC1)|0.134|0.087|1.55|
|absdiffScalarDouble::BinaryOpTest::(640x480, 8SC1)|0.140|0.088|1.58|
|absdiffScalarDouble::BinaryOpTest::(640x480, 16SC1)|0.122|0.087|1.41|
|absdiffScalarDouble::BinaryOpTest::(640x480, 32SC1)|0.093|0.091|1.02|
|absdiffScalarDouble::BinaryOpTest::(640x480, 32FC1)|0.096|0.091|1.05|
|absdiffScalarDouble::BinaryOpTest::(640x480, 16SC2)|0.270|0.190|1.42|
|absdiffScalarDouble::BinaryOpTest::(640x480, 8UC3)|0.413|0.291|1.42|
|absdiffScalarDouble::BinaryOpTest::(640x480, 16SC3)|0.425|0.331|1.28|
|absdiffScalarDouble::BinaryOpTest::(640x480, 8UC4)|0.483|0.410|1.18|
|absdiffScalarDouble::BinaryOpTest::(640x480, 16SC4)|0.544|0.358|1.52|
|absdiffScalarDouble::BinaryOpTest::(1280x720, 8UC1)|0.363|0.307|1.18|
|absdiffScalarDouble::BinaryOpTest::(1280x720, 8SC1)|0.418|0.281|1.49|
|absdiffScalarDouble::BinaryOpTest::(1280x720, 16SC1)|0.364|0.325|1.12|
|absdiffScalarDouble::BinaryOpTest::(1280x720, 32SC1)|0.283|0.285|0.99|
|absdiffScalarDouble::BinaryOpTest::(1280x720, 32FC1)|0.292|0.279|1.04|
|absdiffScalarDouble::BinaryOpTest::(1280x720, 16SC2)|0.815|0.537|1.52|
|absdiffScalarDouble::BinaryOpTest::(1280x720, 8UC3)|1.235|0.898|1.38|
|absdiffScalarDouble::BinaryOpTest::(1280x720, 16SC3)|1.300|0.957|1.36|
|absdiffScalarDouble::BinaryOpTest::(1280x720, 8UC4)|1.431|1.037|1.38|
|absdiffScalarDouble::BinaryOpTest::(1280x720, 16SC4)|1.677|1.249|1.34|
|absdiffScalarDouble::BinaryOpTest::(1920x1080, 8UC1)|0.804|0.584|1.38|
|absdiffScalarDouble::BinaryOpTest::(1920x1080, 8SC1)|0.935|0.583|1.61|
|absdiffScalarDouble::BinaryOpTest::(1920x1080, 16SC1)|0.828|0.582|1.42|
|absdiffScalarDouble::BinaryOpTest::(1920x1080, 32SC1)|0.832|0.877|0.95|
|absdiffScalarDouble::BinaryOpTest::(1920x1080, 32FC1)|0.841|0.861|0.98|
|absdiffScalarDouble::BinaryOpTest::(1920x1080, 16SC2)|1.906|1.412|1.35|
|absdiffScalarDouble::BinaryOpTest::(1920x1080, 8UC3)|2.787|2.037|1.37|
|absdiffScalarDouble::BinaryOpTest::(1920x1080, 16SC3)|3.028|2.554|1.19|
|absdiffScalarDouble::BinaryOpTest::(1920x1080, 8UC4)|3.249|2.727|1.19|
|absdiffScalarDouble::BinaryOpTest::(1920x1080, 16SC4)|3.855|2.874|1.34|
|absdiffScalarSameType::BinaryOpTest::(640x480, 8UC1)|0.018|0.013|1.42|
|absdiffScalarSameType::BinaryOpTest::(640x480, 8SC1)|0.139|0.089|1.56|
|absdiffScalarSameType::BinaryOpTest::(640x480, 16SC1)|0.136|0.100|1.36|
|absdiffScalarSameType::BinaryOpTest::(640x480, 32SC1)|0.100|0.091|1.09|
|absdiffScalarSameType::BinaryOpTest::(640x480, 32FC1)|0.094|0.091|1.03|
|absdiffScalarSameType::BinaryOpTest::(640x480, 16SC2)|0.246|0.179|1.38|
|absdiffScalarSameType::BinaryOpTest::(640x480, 8UC3)|0.079|0.069|1.14|
|absdiffScalarSameType::BinaryOpTest::(640x480, 16SC3)|0.426|0.316|1.35|
|absdiffScalarSameType::BinaryOpTest::(640x480, 8UC4)|0.093|0.093|1.00|
|absdiffScalarSameType::BinaryOpTest::(640x480, 16SC4)|0.485|0.364|1.33|
|absdiffScalarSameType::BinaryOpTest::(1280x720, 8UC1)|0.074|0.068|1.09|
|absdiffScalarSameType::BinaryOpTest::(1280x720, 8SC1)|0.369|0.278|1.33|
|absdiffScalarSameType::BinaryOpTest::(1280x720, 16SC1)|0.416|0.262|1.59|
|absdiffScalarSameType::BinaryOpTest::(1280x720, 32SC1)|0.296|0.284|1.04|
|absdiffScalarSameType::BinaryOpTest::(1280x720, 32FC1)|0.279|0.279|1.00|
|absdiffScalarSameType::BinaryOpTest::(1280x720, 16SC2)|0.738|0.516|1.43|
|absdiffScalarSameType::BinaryOpTest::(1280x720, 8UC3)|0.238|0.215|1.11|
|absdiffScalarSameType::BinaryOpTest::(1280x720, 16SC3)|1.296|0.896|1.45|
|absdiffScalarSameType::BinaryOpTest::(1280x720, 8UC4)|0.302|0.279|1.08|
|absdiffScalarSameType::BinaryOpTest::(1280x720, 16SC4)|1.538|1.139|1.35|
|absdiffScalarSameType::BinaryOpTest::(1920x1080, 8UC1)|0.156|0.153|1.01|
|absdiffScalarSameType::BinaryOpTest::(1920x1080, 8SC1)|0.779|0.594|1.31|
|absdiffScalarSameType::BinaryOpTest::(1920x1080, 16SC1)|0.901|0.646|1.39|
|absdiffScalarSameType::BinaryOpTest::(1920x1080, 32SC1)|0.852|0.879|0.97|
|absdiffScalarSameType::BinaryOpTest::(1920x1080, 32FC1)|0.810|0.859|0.94|
|absdiffScalarSameType::BinaryOpTest::(1920x1080, 16SC2)|1.736|1.312|1.32|
|absdiffScalarSameType::BinaryOpTest::(1920x1080, 8UC3)|0.565|0.555|1.02|
|absdiffScalarSameType::BinaryOpTest::(1920x1080, 16SC3)|3.079|2.529|1.22|
|absdiffScalarSameType::BinaryOpTest::(1920x1080, 8UC4)|0.838|0.845|0.99|
|absdiffScalarSameType::BinaryOpTest::(1920x1080, 16SC4)|3.465|2.863|1.21|
|add::BinaryOpTest::(640x480, 8UC1)|0.011|0.011|0.94|
|add::BinaryOpTest::(640x480, 8SC1)|0.011|0.011|0.95|
|add::BinaryOpTest::(640x480, 16SC1)|0.064|0.062|1.04|
|add::BinaryOpTest::(640x480, 32SC1)|0.133|0.133|1.00|
|add::BinaryOpTest::(640x480, 32FC1)|0.134|0.133|1.00|
|add::BinaryOpTest::(640x480, 16SC2)|0.132|0.133|0.99|
|add::BinaryOpTest::(640x480, 8UC3)|0.100|0.100|1.01|
|add::BinaryOpTest::(640x480, 16SC3)|0.200|0.200|1.00|
|add::BinaryOpTest::(640x480, 8UC4)|0.134|0.133|1.00|
|add::BinaryOpTest::(640x480, 16SC4)|0.271|0.269|1.01|
|add::BinaryOpTest::(1280x720, 8UC1)|0.100|0.099|1.01|
|add::BinaryOpTest::(1280x720, 8SC1)|0.099|0.101|0.98|
|add::BinaryOpTest::(1280x720, 16SC1)|0.200|0.201|0.99|
|add::BinaryOpTest::(1280x720, 32SC1)|0.418|0.420|1.00|
|add::BinaryOpTest::(1280x720, 32FC1)|0.425|0.446|0.95|
|add::BinaryOpTest::(1280x720, 16SC2)|0.421|0.424|0.99|
|add::BinaryOpTest::(1280x720, 8UC3)|0.304|0.305|1.00|
|add::BinaryOpTest::(1280x720, 16SC3)|0.726|0.741|0.98|
|add::BinaryOpTest::(1280x720, 8UC4)|0.424|0.424|1.00|
|add::BinaryOpTest::(1280x720, 16SC4)|1.077|1.123|0.96|
|add::BinaryOpTest::(1920x1080, 8UC1)|0.225|0.226|1.00|
|add::BinaryOpTest::(1920x1080, 8SC1)|0.225|0.227|0.99|
|add::BinaryOpTest::(1920x1080, 16SC1)|0.479|0.477|1.00|
|add::BinaryOpTest::(1920x1080, 32SC1)|1.278|1.324|0.97|
|add::BinaryOpTest::(1920x1080, 32FC1)|1.298|1.438|0.90|
|add::BinaryOpTest::(1920x1080, 16SC2)|1.267|1.313|0.96|
|add::BinaryOpTest::(1920x1080, 8UC3)|0.841|0.851|0.99|
|add::BinaryOpTest::(1920x1080, 16SC3)|2.055|2.115|0.97|
|add::BinaryOpTest::(1920x1080, 8UC4)|1.271|1.313|0.97|
|add::BinaryOpTest::(1920x1080, 16SC4)|2.848|2.963|0.96|
|addScalarDouble::BinaryOpTest::(640x480, 8UC1)|0.126|0.090|1.40|
|addScalarDouble::BinaryOpTest::(640x480, 8SC1)|0.116|0.088|1.32|
|addScalarDouble::BinaryOpTest::(640x480, 16SC1)|0.128|0.095|1.34|
|addScalarDouble::BinaryOpTest::(640x480, 32SC1)|0.094|0.091|1.04|
|addScalarDouble::BinaryOpTest::(640x480, 32FC1)|0.093|0.090|1.03|
|addScalarDouble::BinaryOpTest::(640x480, 16SC2)|0.229|0.164|1.40|
|addScalarDouble::BinaryOpTest::(640x480, 8UC3)|0.386|0.270|1.43|
|addScalarDouble::BinaryOpTest::(640x480, 16SC3)|0.391|0.289|1.35|
|addScalarDouble::BinaryOpTest::(640x480, 8UC4)|0.490|0.372|1.32|
|addScalarDouble::BinaryOpTest::(640x480, 16SC4)|0.445|0.342|1.30|
|addScalarDouble::BinaryOpTest::(1280x720, 8UC1)|0.371|0.265|1.40|
|addScalarDouble::BinaryOpTest::(1280x720, 8SC1)|0.343|0.247|1.39|
|addScalarDouble::BinaryOpTest::(1280x720, 16SC1)|0.378|0.274|1.38|
|addScalarDouble::BinaryOpTest::(1280x720, 32SC1)|0.284|0.281|1.01|
|addScalarDouble::BinaryOpTest::(1280x720, 32FC1)|0.279|0.278|1.00|
|addScalarDouble::BinaryOpTest::(1280x720, 16SC2)|0.669|0.521|1.28|
|addScalarDouble::BinaryOpTest::(1280x720, 8UC3)|1.134|0.799|1.42|
|addScalarDouble::BinaryOpTest::(1280x720, 16SC3)|1.198|0.912|1.31|
|addScalarDouble::BinaryOpTest::(1280x720, 8UC4)|1.475|1.077|1.37|
|addScalarDouble::BinaryOpTest::(1280x720, 16SC4)|1.419|1.068|1.33|
|addScalarDouble::BinaryOpTest::(1920x1080, 8UC1)|0.836|0.600|1.39|
|addScalarDouble::BinaryOpTest::(1920x1080, 8SC1)|0.761|0.641|1.19|
|addScalarDouble::BinaryOpTest::(1920x1080, 16SC1)|0.863|0.583|1.48|
|addScalarDouble::BinaryOpTest::(1920x1080, 32SC1)|0.859|0.874|0.98|
|addScalarDouble::BinaryOpTest::(1920x1080, 32FC1)|0.820|0.862|0.95|
|addScalarDouble::BinaryOpTest::(1920x1080, 16SC2)|1.635|1.225|1.33|
|addScalarDouble::BinaryOpTest::(1920x1080, 8UC3)|2.598|2.109|1.23|
|addScalarDouble::BinaryOpTest::(1920x1080, 16SC3)|2.823|2.131|1.32|
|addScalarDouble::BinaryOpTest::(1920x1080, 8UC4)|3.310|2.291|1.44|
|addScalarDouble::BinaryOpTest::(1920x1080, 16SC4)|3.325|2.947|1.13|
|addScalarSameType::BinaryOpTest::(640x480, 8UC1)|0.018|0.013|1.35|
|addScalarSameType::BinaryOpTest::(640x480, 8SC1)|0.115|0.084|1.37|
|addScalarSameType::BinaryOpTest::(640x480, 16SC1)|0.129|0.084|1.53|
|addScalarSameType::BinaryOpTest::(640x480, 32SC1)|0.096|0.091|1.06|
|addScalarSameType::BinaryOpTest::(640x480, 32FC1)|0.093|0.090|1.03|
|addScalarSameType::BinaryOpTest::(640x480, 16SC2)|0.232|0.188|1.24|
|addScalarSameType::BinaryOpTest::(640x480, 8UC3)|0.075|0.071|1.07|
|addScalarSameType::BinaryOpTest::(640x480, 16SC3)|0.400|0.272|1.47|
|addScalarSameType::BinaryOpTest::(640x480, 8UC4)|0.095|0.092|1.04|
|addScalarSameType::BinaryOpTest::(640x480, 16SC4)|0.451|0.330|1.36|
|addScalarSameType::BinaryOpTest::(1280x720, 8UC1)|0.071|0.068|1.04|
|addScalarSameType::BinaryOpTest::(1280x720, 8SC1)|0.347|0.279|1.24|
|addScalarSameType::BinaryOpTest::(1280x720, 16SC1)|0.385|0.265|1.45|
…